ブロック構文の追加

+++ main.c
@@ -44,7 +44,7 @@ void runtest(){
     expect(__LINE__, 0, vec->len);
 
     for(int i=0;i<100;++i)
-        vec_push(vec, (void*) i);
+        vec_push(vec, (void*) (long) i);
 
     expect(__LINE__, 100, vec->len);
     expect(__LINE__,  0, (long)vec->data[0]);
